phase_3/requirements.txt
groq>=1.0.0
python-dotenv==1.0.0
sentence-transformers==2.3.1
faiss-cpu==1.13.2
numpy==1.26.3
pyyaml==6.0.1
streamlit==1.40.2
